requirement.txt
pymongo
pandas
numpy
pymongo[srv]
scikit-learn
python-dotenv
certifi
PyYAML
dill
setuptools
scipy
imblearn

-e .